+++ rpcs3/Emu/RSX/GL/rsx_gl_texture.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "OpenGL.h"
 
 namespace rsx
